More cleanup
• Unify type definitions
• Simplify layout
• Reuse existing layout components
• Fix JS directory structure
