+++ assets/js/@types/gluggi.d.ts
@@ -2,10 +2,18 @@ export module Gluggi
 {
     export interface Component
     {
-        type: string;
         key: string;
+        type: string;
         name: string;
         hidden: boolean;
+        error: string|null;
+        active: boolean;
+    }
+
+    export interface Type
+    {
+        title: string;
+        components: Component[];
     }
 
     export interface Dependencies
