+++ assets/js/components/ComponentLink.tsx
@@ -1,7 +1,7 @@
 import {h, JSX, RenderableProps} from "preact";
 import {Gluggi} from "../@types/gluggi";
-import {GluggiRouter} from "./GluggiRouter";
-import {icon} from "./icon";
+import {GluggiRouter} from "../lib/GluggiRouter";
+import {icon} from "../lib/icon";
 
 // @ts-ignore
 import arrowIcon from "../../icon/active.svg";
@@ -21,21 +21,20 @@ export interface ComponentLinkProps
 export function ComponentLink (props: RenderableProps<ComponentLinkProps>): JSX.Element
 {
     let component = props.component;
-    let isActive = true === props.active;
 
     return (
-        <a class={`gluggi-component-link ${isActive ? " gluggi-active" : ""}`} href={props.router.component(component)}>
-            {isActive && (
+        <a class={`gluggi-component-link ${props.component.active ? " gluggi-active" : ""}`} href={props.router.component(component)}>
+            {props.component.active && (
                 icon(arrowIcon)
             )}
             <span class="gluggi-component-name">
-                    {(true !== props.shortName) && (
-                        <span class="gluggi-component-type">{component.type}</span>
-                    )}
+                {(true !== props.shortName) && (
+                    <span class="gluggi-component-type">{component.type}</span>
+                )}
                 {` ${component.name}`}
                 {props.children}
                 {"page" === component.type ? icon(externalIcon, "isolated view") : ""}
-                </span>
+            </span>
         </a>
     );
 }
