+++ assets/scss/atom/_data-container.scss
@@ -1,0 +1,3 @@
+.gluggi-data-container {
+    display: none;
+}
