+++ assets/scss/atom/_stage.scss
@@ -154,7 +154,7 @@
             padding: 50px;
         }
 
-        &.is-fullscreen {
+        &.gluggi-is-fullscreen {
             padding: 0;
             background-color: $content-bg;
