+++ assets/scss/gluggi.scss
@@ -6,6 +6,7 @@
 // Atoms
 @import "atom/component-link";
 @import "atom/components";
+@import "atom/data-container";
 @import "atom/empty";
 @import "atom/search";
 @import "atom/sidebar-content-data";
