+++ composer.json
@@ -12,8 +12,8 @@
     ],
     "require": {
         "php": "^7.3",
+        "ext-json": "*",
         "becklyn/assets-bundle": "^2.5",
-        "becklyn/rad-bundle": "^6.1",
         "symfony/dependency-injection": "^4.1",
         "symfony/finder": "^4.1",
         "symfony/framework-bundle": "^4.1",
@@ -23,6 +23,7 @@
         "symfony/twig-bundle": "^4.1",
         "symfony/yaml": "^4.1",
         "twig/twig": "^2.10"
+
     },
     "require-dev": {
         "becklyn/php-cs": "^1.5",
@@ -30,7 +31,8 @@
         "symfony/browser-kit": "^4.1",
         "symfony/css-selector": "^4.1",
         "symfony/phpunit-bridge": "^4.1",
-        "symfony/twig-bundle": "^4.1"
+        "symfony/twig-bundle": "^4.1",
+        "symfony/var-dumper": "^4.1"
     },
     "config": {
         "sort-packages": true
