--- tests/Component/ComponentLoaderTest.php
@@ -1,40 +1,0 @@
-<?php
-
-namespace Tests\Becklyn\GluggiBundle\Component;
-
-use Becklyn\GluggiBundle\Type\TypeFactory;
-use Becklyn\GluggiBundle\Data\ComponentType;
-use PHPUnit\Framework\TestCase;
-
-
-/**
- *
- */
-class ComponentLoaderTest extends TestCase
-{
-    private function loadAtoms ()
-    {
-        $loader = new TypeFactory(__DIR__ . "/../templates/layout", "layout");
-        $type = self::getMockBuilder(ComponentType::class)
-            ->disableOriginalConstructor()
-            ->getMock();
-
-        $type
-            ->method("getDirectory")
-            ->willReturn("atom");
-
-        return $loader->loadComponents($type);
-    }
-
-
-
-    public function testViewDirectory ()
-    {
-        $components = $this->loadAtoms();
-
-        self::assertArrayHasKey("_hidden", $components);
-        self::assertArrayHasKey("example", $components);
-        self::assertArrayHasKey("standalone_test", $components);
-        self::assertCount(3, $components);
-    }
-}
